feat: generic equipment card styles
