+++ src/pages/charsheet.astro
@@ -4,6 +4,7 @@ import Layout from '../layouts/layout.astro';
 import { Charsheet as CharsheetComponent } from '../components/charsheet';
 import '../styles/charsheet.css';
 import '../styles/tabs.css';
+import '../styles/item-card.css';
 import '../styles/armor/index.css';
 import '../styles/bar/index.css';
 import '../styles/skills.css';
