+++ src/styles/item-card.css
@@ -1,0 +1,46 @@
+/* === Shared Item Card Base (gear, armor, weapons, cyber) === */
+.item-card {
+  background: var(--code-bg);
+  border: 1px solid var(--border-light);
+  border-radius: var(--radius-lg);
+  padding: var(--padding-panel);
+  margin-bottom: var(--space-8);
+  cursor: pointer;
+}
+
+.item-card:hover {
+  border-color: var(--accent);
+}
+
+.item-card.selected {
+  border-color: var(--accent);
+  box-shadow: 0 0 8px var(--accent-muted);
+}
+
+.item-card.highlighted {
+  background: var(--accent-subtle);
+  border-color: var(--accent);
+}
+
+.item-card.item-card-accent {
+  border-left: 3px solid var(--accent);
+}
+
+.item-card h4 {
+  margin: 0;
+  font-size: var(--font-ui-lg);
+  color: var(--accent);
+}
+
+.item-meta {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  gap: 2px;
+}
+
+.item-avail {
+  font-size: var(--font-ui-sm);
+  font-weight: bold;
+  color: var(--fg-soft);
+}
